Create ptrauth.c (FreeBSD) - Integration required.
This file is from FreeBSD. There is code that accompanies it that had yet to be imported. Integration into the OpenBSD specific arm64 codebase here will be required. In addition, edits to the aformentioned file may be required as well. In short, as of this commit, there is no valid pointer authentication mechanism integrated into the working system...