Commented out error-prone conditionals
Conditionals in this file were not agreeing with the OpenBSD build system and were throwing errors. I believe this is likely because they are calling upon values which wold have been declared in the above included file which we commented out. While nice to have this different build configurations, defining these variables in the include file which could be effectively included for use on this system is not the highest priority, nor does it appear to be mission critical.